Moved HidraConstants into separate file refs #219
