+++ pyrs/core/reduction_manager.py
@@ -2,14 +2,15 @@
 import os
 import random
 import numpy as np
-from pyrs.utilities import checkdatatypes
 from pyrs.core import workspaces
-from pyrs.utilities import calibration_file_io
+from pyrs.core import instrument_geometry
 from pyrs.core import mask_util
 from pyrs.core import reduce_hb2b_mtd
 from pyrs.core import reduce_hb2b_pyrs
-from pyrs.utilities.rs_project_file import HidraProjectFile, HidraProjectFileMode, HidraConstants
-from pyrs.core import instrument_geometry
+from pyrs.projectfile import HidraConstants
+from pyrs.utilities import calibration_file_io
+from pyrs.utilities import checkdatatypes
+from pyrs.utilities.rs_project_file import HidraProjectFile, HidraProjectFileMode
 
 
 class HB2BReductionManager(object):
