+++ pyrs/projectfile/__init__.py
@@ -1,0 +1,4 @@
+# flake8: noqa
+from .constants import HidraConstants
+
+__all__ = ['HidraConstants']
